% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotBasis.R
\name{rotBasis}
\alias{rotBasis}
\title{Generate a complete orthonormal basis including vectors orthogonal to a given vector}
\usage{
rotBasis(x)
}
\arguments{
\item{x}{A vector of length \eqn{d} whose direction will be the first column 
of the resulting orthogonal matrix.}
}
\value{
A \eqn{d}-by-\eqn{d} orthogonal matrix, 
of which the first column being the direction of \code{x}, 
a complete orthonormal basis consists of columns of the output matrix.
}
\description{
With its first column being the given vector divided by its norm, 
generate an orthogonal matrix by rotating the identity matrix.
}
\details{
Let \eqn{e_k} be the \eqn{k}th column of a \eqn{d}-dimensional identity matrix, 
for \eqn{k=1,\dots,d}. 
With \eqn{e_x = ( x - (x^\top e_1) e_1 ) / \| x - (x^\top e_1) e_1 \|} such that 
\eqn{x = (x^\top e_1) e_1 + (x^\top e_x) e_x}, 
let \eqn{Q} be the Q component of the QR decomposition of 
\eqn{( e_1\ e_x\ e_2\ \cdots\ e_{d-1})}. 
Then the output orthogonal matrix is given by \eqn{Q D Q^\top}. 
Here, \eqn{D} is a block diagonal matrix with two diagonal blocks.
The first diagonal block is a \eqn{2}-by-\eqn{2} matrix with entries given by 
\eqn{D_{11} = x^\top e_1}, \eqn{D_{12} = -x^\top e_x}, 
\eqn{D_{21} = x^\top e_x}, and \eqn{D_{22} = x^\top e_1}. 
The second diagonal block is a \eqn{(d-2)}-dimensional identity matrix.
}
